yzs/yzs/apps/requirements.txt
amqp==5.0.2
billiard==3.6.3.0
celery==5.0.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
dappy-drf-extensions==0.4.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.17
django-ckedit==1.0
django-cors-headers==3.5.0
django-crontab==0.7.1
django-filter==2.4.0
django-haystack==3.0
django-import-export==2.4.0
django-markup==1.5
django-redis==4.12.1
django-reversion==3.0.8
djangorestframework==3.12.2
djangorestframework-apidoc==0.0.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-haystack==1.8.9
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==3.1.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
prompt-toolkit==3.0.8
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
